# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NetworkSim API",
    "description": "Network Topology Simulator with Chaos Engineering capabilities.\n\n## Features\n- Create and manage network topologies\n- Deploy topologies to Kubernetes\n- Apply chaos engineering (delay, loss, partition, etc.)\n- Monitor network metrics and events\n- Run diagnostic tests",
    "contact": {
      "name": "NetworkSim Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/scenarios/{id}": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "operationId": "get_scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scenario ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario"
          }
        }
      },
      "put": {
        "tags": [
          "scenarios"
        ],
        "operationId": "update_scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scenario ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scenario updated"
          }
        }
      },
      "delete": {
        "tags": [
          "scenarios"
        ],
        "operationId": "delete_scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scenario ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario deleted"
          }
        }
      }
    },
    "/api/scenarios/{id}/run": {
      "post": {
        "tags": [
          "scenarios"
        ],
        "operationId": "run_scenario",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Scenario ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scenario execution started"
          }
        }
      }
    },
...
    "/api/topologies/{topology_id}/scenarios": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "operationId": "list_scenarios",
        "parameters": [
          {
            "name": "topology_id",
            "in": "path",
            "description": "Topology ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scenarios"
          }
        }
      },
      "post": {
        "tags": [
          "scenarios"
        ],
        "operationId": "create_scenario",
        "parameters": [
          {
            "name": "topology_id",
            "in": "path",
            "description": "Topology ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScenarioRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Scenario created"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChaosType": {
        "type": "string",
        "description": "Types of chaos conditions that can be applied",
        "enum": [
          "delay",
          "loss",
          "bandwidth",
          "corrupt",
          "duplicate",
          "partition",
          "stress-cpu",
          "pod-kill",
          "io-delay",
          "http-abort"
        ]
      },
...
      "CreateScenarioRequest": {
        "type": "object",
        "description": "Request to create a new scenario for a topology.",
        "required": [
          "name",
          "total_duration",
          "steps"
        ],
        "properties": {
          "description": {
            "type": "string",
            "example": "Optional description",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "Baseline failure test"
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenarioStep"
            }
          },
          "total_duration": {
            "type": "integer",
            "format": "int64",
            "example": 60
          }
        }
      },
...
      "ScenarioStep": {
        "type": "object",
        "description": "Single step inside a scenario, describing a chaos action to apply.",
        "required": [
          "id",
          "type",
          "sourceNodeId",
          "startAt",
          "duration",
          "params",
          "laneId"
        ],
        "properties": {
          "duration": {
            "type": "number",
            "format": "double",
            "description": "Duration in seconds",
            "example": 10.0
          },
          "id": {
            "type": "string",
            "example": "step-1"
          },
          "laneId": {
            "type": "string",
            "example": "lane-1"
          },
          "params": {
            "description": "Step-specific parameters as JSON"
          },
          "sourceNodeId": {
            "type": "string",
            "description": "Source node id for the action",
            "example": "node-1"
          },
          "startAt": {
            "type": "number",
            "format": "double",
            "description": "Start time (seconds) relative to scenario start",
            "example": 0.0
          },
          "targetNodeId": {
            "type": "string",
            "description": "Optional target node id for the action",
            "example": "node-2",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/ChaosType"
          }
        }
      },
...
      "UpdateScenarioRequest": {
        "type": "object",
        "description": "Partial update for an existing scenario.",
        "properties": {
          "description": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "example": "Updated scenario name",
            "nullable": true
          },
          "steps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ScenarioStep"
            },
            "nullable": true
          },
          "total_duration": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          }
        }
      },
...
    }
  },
...
}